pyproject.toml
[tool.poetry]
name = "feed_bot"
version = "0.1.0"
description = "A discord bot for posting rss feeds and read-only json apis from websites"
authors = ["Christopher Dixon <chrisdixononcode@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
"discord.py" = "^2.3.2"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
requests-random-user-agent = "^2023.2.15"
motor = "^3.3.2"

[tool.poetry.scripts]
bot = 'feed_bot.bot:main'

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
